[FEATURE] UI: multiline when lexing incomplete

+++ srcs/history/history_arrow.c
@@ -15,7 +15,7 @@
 BOOL	history_exec_arrow_right(t_input *input)
 {
 	(void)input;
-	if (((input->cpos.cp_col + (input->offset_line  * input->ts.ws_col) - input->offset_col)) < input->str->len)
+	if (((size_t)(input->cpos.cp_col + (input->offset_line  * input->ts.ws_col) - input->offset_col)) < input->str->len)
 	{
 		if (input->cpos.cp_col + 1 == input->ts.ws_col)
 			input->offset_line += 1;
